Refactor vscode extension
