+++ editors/code/src/commands.ts
@@ -1,0 +1,17 @@
+import * as applySourceChange from './commands/apply_source_change';
+import * as extendSelection from './commands/extend_selection';
+import * as joinLines from './commands/join_lines';
+import * as matchingBrace from './commands/matching_brace';
+import * as parentModule from './commands/parent_module';
+import * as runnables from './commands/runnables';
+import * as syntaxTree from './commands/syntaxTree';
+
+export {
+    applySourceChange,
+    extendSelection,
+    joinLines,
+    matchingBrace,
+    parentModule,
+    runnables,
+    syntaxTree
+}
