+++ editors/code/src/events.ts
@@ -1,0 +1,7 @@
+import * as changeActiveTextEditor from './events/change_active_text_editor'
+import * as changeTextDocument from './events/change_text_document';
+
+export {
+    changeActiveTextEditor,
+    changeTextDocument
+}
